feat: Package using Angular Package Format
