+++ src/public_api.ts
@@ -1,0 +1,2 @@
+export * from './lib/theme-config';
+export * from './lib/dark-switch';
